fix edit site form autocomplete
