+++ packages/frinx-gamma/src/components/forms/vpn-site-form.tsx
@@ -18,7 +18,6 @@ import CustomerLocationForm from './customer-location-form';
 import SiteDeviceForm from './site-device-form';
 import Autocomplete from '../autocomplete/autocomplete';
 import unwrap from '../../helpers/unwrap';
-import Autocomplete2, { Item } from '../autocomplete-2/autocomplete-2';
 
 type Props = {
   mode: 'add' | 'edit';
@@ -44,7 +43,7 @@ const getDefaultSiteDevice = (): SiteDevice => ({
   managementIP: '',
 });
 
-const VpnSiteForm: FC<Props> = ({ mode, site, sites, qosProfiles, onSubmit, onCancel, onSiteChange }) => {
+const VpnSiteForm: FC<Props> = ({ site, qosProfiles, onSubmit, onCancel }) => {
   const [siteState, setSiteState] = useState(site);
   const [customerLocationsForm, setCustomerLocationsForm] = useState<CustomerLocation | null>(null); // eslint-disable-line @typescript-eslint/no-unused-vars
 
@@ -112,34 +111,8 @@ const VpnSiteForm: FC<Props> = ({ mode, site, sites, qosProfiles, onSubmit, onCa
     });
   };
 
-  const handleSiteItemChange = (item?: Item | null) => {
-    if (onSiteChange) {
-      const [newSite] = sites.filter((s) => s.siteId === item?.value);
-      onSiteChange(newSite);
-    }
-  };
-
-  const siteItems =
-    mode === 'edit'
-      ? sites
-          .map((s) => ({
-            value: unwrap(s.siteId),
-            label: unwrap(s.siteId),
-          }))
-          .filter((s) => s.value !== siteState.siteId)
-      : [];
-
-  const [selectedSiteItem] = siteItems.filter((item) => item.value === siteState.siteId);
-
   return (
     <form onSubmit={handleSubmit}>
-      {mode === 'edit' && (
-        <FormControl id="siteId" my={6}>
-          <FormLabel>Site ID</FormLabel>
-          {/* <Autocomplete items={vpnIds} selectedItem={serviceState.vpnId || ''} onChange={handleVpnIdChange} /> */}
-          <Autocomplete2 items={siteItems} selectedItem={selectedSiteItem} onChange={handleSiteItemChange} />
-        </FormControl>
-      )}
       <FormControl id="customer-locations" my={6}>
         <FormLabel>Cutomer Locations</FormLabel>
         <Box
