+++ packages/frinx-gamma/src/index.ts
@@ -2,7 +2,7 @@ export { default as Main } from './pages/main/main';
 export { default as CreateVpnService } from './pages/create-vpn-service/create-vpn-service';
 export { default as EditVpnService } from './pages/edit-vpn-service/edit-vpn-service';
 export { default as CreateVpnSite } from './pages/create-vpn-site/create-vpn-site';
-export { default as EditVpnSite } from './pages/edit-vpn-service/edit-vpn-service';
+export { default as EditVpnSite } from './pages/edit-vpn-site/edit-vpn-site';
 export { default as CreateSiteNetAccess } from './pages/create-site-network-access/create-site-network-access';
 export { default as EditSiteNetAccess } from './pages/edit-site-network-access/edit-site-network-access';
 export { getUnistoreApiProvider } from './unistore-api-provider';
